core/pta/remote_attestation: add build configuration for remote attestation PTA
This commit adds the necessary build configuration for the remote attestation
PTA. It includes adding the remote_attestation directory to the build in
sub.mk and introducing the CFG_VERAISON_ATTESTATION_PTA option in config.mk.
When enabled, this option embeds the remote attestation PTA service and forces
the CFG_QCBOR option to ensure the required dependencies are included.

Signed-off-by: Yuichi Sugiyama <yuichis@ricsec.co.jp>